+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Version 1.0.4
 
   * Add notice board option to getting My.SCOUT link for a member (pass :notice as the parameter)
+  * Model.require\_ability\_to now only checks subscription level for youth sections
 
 ## Version 1.0.3